refactor(agents): compose useAgentChat from focused sub-hooks
Split the remaining responsibilities of useAgentChat into three
sub-hooks so the composition root reads as wiring:

- useAgentSessionSync: busy-elsewhere entry, the sync-probe poll loop,
  and cadence selection (comments and race guards moved verbatim)
- useDraftSessionCreation: create-session mutation, optimistic pending
  draft message, and first-send handoff to the new runtime chat
- useAgentSessionHistory: rewind (truncate) and branch mutations plus
  dropped-tool-state release

The GraphQL operations move with their hooks (Relay artifacts renamed
accordingly), AgentChatOperationError moves to its own module (re-
exported for existing consumers), and isRequestActive moves to
agent/chat/chatStatus for shared use. No behavior change.
